+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * Argument `split_by` to make multiple plot given a variable in `sam_data` slot(function `ggvenn_pq`)  
 * Argument `seq_names` in `asv2otu` function allow to clusterize sequences from a character vector of DNA.
+* Add a `blast_pq` function to blast the sequences of the `ref_seq` slot against a custom database
+* Add a `filter_asv_blast` function to filter ASV in phyloseq dataset using blast against a custom database
 
 # MiscMetabar 0.3